fec: used typecasts to fix compiler warnings

+++ gr-fec/lib/async_encoder_impl.cc
@@ -117,9 +117,9 @@ namespace gr {
         nbits_out = d_encoder->get_output_size();
         } else {
         nblocks = nbits_in / d_encoder->get_input_size();
-        if( nblocks * d_encoder->get_input_size() != nbits_in ){
+        if( (int)(nblocks * d_encoder->get_input_size()) != nbits_in ){
             printf("nblocks: %d, in_block_size: %d, got_input_size: %d\n",
-                nblocks, d_encoder->get_input_size(), nbits_in);
+                (int)nblocks, (int)(d_encoder->get_input_size()), nbits_in);
             throw std::runtime_error("input does not divide into code block size!");
             }
         nbits_out = nblocks * d_encoder->get_output_size();
